enable manual tests with cec-monitor using exit timeout
