+++ package.json
@@ -40,7 +40,6 @@
   },
   "dependencies": {
     "@senzil/cec-monitor": "^2.0.0-release.79",
-    "cec-promise": "^0.2.0",
     "common-display-module": "git://github.com/Rise-Vision/common-display-module.git#v1.5.5",
     "node-ipc": "^9.1.1",
     "rise-common-electron": "git://github.com/Rise-Vision/rise-common-electron.git#v2.2.3"
